Refactor ens lookup and ensTlds into composable
